+++ packages/messaging/src/websockets/events/payload.ts
@@ -4,6 +4,14 @@ import type {
   IEventSocialCreationFailedPayload,
   IEventSocialDeletedPayload,
   IEventSocialDeletionFailedPayload,
+  IFallbackChangeEventStatePayload,
+  IFallbackCreateEventPayload,
+  IFallbackCreateTagPayload,
+  IFallbackDeleteEventPayload,
+  IFallbackDeleteTagPayload,
+  IFallbackManageRequirementsPayload,
+  IFallbackUpdateEventPayload,
+  IFallbackUpdateTagPayload,
 } from '../../index.js';
 
 export interface IEventCreatedWebsocketPayload
@@ -17,4 +25,29 @@ export interface IEventDeletedWebsocketPayload
     IEmitterPayload {}
 export interface IEventDeletionFailedWebsocketPayload
   extends IEventSocialDeletionFailedPayload,
+    IEmitterPayload {}
+
+export interface IFallbackCreateEventWebsocketPayload
+  extends IFallbackCreateEventPayload,
+    IEmitterPayload {}
+export interface IFallbackUpdateEventWebsocketPayload
+  extends IFallbackUpdateEventPayload,
+    IEmitterPayload {}
+export interface IFallbackDeleteEventWebsocketPayload
+  extends IFallbackDeleteEventPayload,
+    IEmitterPayload {}
+export interface IFallbackChangeEventStateWebsocketPayload
+  extends IFallbackChangeEventStatePayload,
+    IEmitterPayload {}
+export interface IFallbackManageRequirementsWebsocketPayload
+  extends IFallbackManageRequirementsPayload,
+    IEmitterPayload {}
+export interface IFallbackCreateTagWebsocketPayload
+  extends IFallbackCreateTagPayload,
+    IEmitterPayload {}
+export interface IFallbackUpdateTagWebsocketPayload
+  extends IFallbackUpdateTagPayload,
+    IEmitterPayload {}
+export interface IFallbackDeleteTagWebsocketPayload
+  extends IFallbackDeleteTagPayload,
     IEmitterPayload {}