+++ packages/post-processors/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 3.0.14
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @volontariapp/messaging@2.9.0
+  - @volontariapp/database@3.3.14
+  - @volontariapp/outbox@0.9.32
+
 ## 3.0.13
 
 ### Patch Changes